Refactor Kafka processors
