+++ hazelcast-jet-kafka/src/main/java/com/hazelcast/jet/processors/package-info.java
@@ -15,6 +15,6 @@
  */
 
 /**
- * Kafka reader/writer support for Hazelcast Jet
+ * Kafka reader/writer support for Hazelcast Jet.
  */
-package com.hazelcast.jet.connector.kafka;
+package com.hazelcast.jet.processors;
